Combat Poke sauvage réalise

+++ lib/attaque.h
@@ -16,7 +16,6 @@ typedef struct attaque{
 attaque * recupAttaque(type t); //retourne un tab de 2 attaque pour le pokemon
 
 //renvoie les dommages d'une attaque 
-int domageattaque(int k,int att_attack,int def_defender,int power);
-
+int domageattaque(float k,int att_attack,int def_defender,int power);
 
 #endif
